requirements.txt
absl-py==1.4.0
addict==2.4.0
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
asttokens==2.2.1
astunparse==1.6.3
attrdict==2.0.1
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
colorama==0.4.6
comm==0.1.3
ConfigArgParse==1.5.3
contourpy==1.0.7
crcmod==1.7
cryptography==41.0.4
cycler==0.11.0
Cython==3.0.2
dash==2.9.3
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.12.0
Flask==2.3.2
flatbuffers==23.3.3
fonttools==4.39.3
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.0
h5py==3.8.0
idna==3.4
imageio==2.28.1
imgaug @ git+https://github.com/aleju/imgaug.git@0101108d4fed06bc5056c4a03e2bcb0216dac326
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipykernel==6.22.0
ipython==8.12.1
ipywidgets==8.0.6
itsdangerous==2.1.2
jax==0.4.8
jedi==0.18.2
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab-widgets==3.0.7
keras==2.12.0
kiwisolver==1.4.4
lazy_loader==0.2
libclang==16.0.0
lit==16.0.2
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.1.0
model-index==0.1.11
mpmath==1.3.0
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.1
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
open3d==0.17.0
opencv-python==4.1.2.30
opendatalab==0.0.10
OpenEXR @ git+https://github.com/jamesbowman/openexrpython.git@91ec8a81a0bad7b30170c66769464cc7ab44ce4b
openmim==0.3.9
openxlab==0.0.26
opt-einsum==3.3.0
ordered-set==4.1.0
oss2==2.17.0
oyaml==1.0
packaging==23.1
pandas==2.0.1
parso==0.8.3
pathlib==1.0.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
plotly==5.14.1
prettytable==3.9.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.19.0
Pygments==2.15.1
pyparsing==3.0.9
pyquaternion==0.9.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.2
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.4.2
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
shapely==2.0.1
six==1.16.0
stack-data==0.6.2
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorboardX==2.6
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
terminaltables==3.1.10
threadpoolctl==3.1.0
tifffile==2023.4.12
tomli==2.0.1
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.3.3
widgetsnbextension==4.0.7
wrapt==1.14.1
yapf==0.40.2
zipp==3.15.0
